backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
cryptography==41.0.7
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
drf-yasg==1.21.7
Pillow==10.1.0
django-cors-headers==4.3.1
django-filter==23.5
faker==20.1.0
setuptools<81